+++ src/main/java/net/frozenblock/lib/integration/impl/EmptyModIntegration.java
@@ -30,7 +30,5 @@ public class EmptyModIntegration extends ModIntegration {
 	}
 
 	@Override
-	public void init() {
-
-	}
+	public void init() {}
 }